gh-3 renamed UI project to .MAUI and dropped .Models from lib project
