+++ RPG Character Sheet.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.7.34031.279
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RPGCharacterSheet", "RPGCharacterSheet\RPGCharacterSheet.csproj", "{3B888450-C9D0-4512-88FB-2EDDB4B5CC0C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RPGCharacterSheet.MAUI", "RPGCharacterSheet.MAUI\RPGCharacterSheet.MAUI.csproj", "{3B888450-C9D0-4512-88FB-2EDDB4B5CC0C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RPGCharacterSheet.Models", "RPGCharacterSheet.Models\RPGCharacterSheet.Models.csproj", "{DA340519-8871-432C-8A14-7F7EC15B0101}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RPGCharacterSheet", "RPGCharacterSheet\RPGCharacterSheet.csproj", "{DA340519-8871-432C-8A14-7F7EC15B0101}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RPGCharacterSheet.Tests", "RPGCharacterSheet.Tests\RPGCharacterSheet.Tests.csproj", "{0D1AB00D-694B-4BD9-BA06-95A12A2F9AFE}"
 EndProject
