+++ RPGCharacterSheet.MAUI/Platforms/Android/AndroidManifest.xml
