+++ RPGCharacterSheet.MAUI/Platforms/Android/MainActivity.cs
