+++ RPGCharacterSheet.MAUI/Platforms/Android/MainApplication.cs
