+++ RPGCharacterSheet.MAUI/Platforms/Android/Resources/values/colors.xml
