+++ RPGCharacterSheet.MAUI/Platforms/MacCatalyst/AppDelegate.cs
