+++ RPGCharacterSheet.MAUI/Platforms/MacCatalyst/Info.plist
