+++ RPGCharacterSheet.MAUI/Platforms/Tizen/tizen-manifest.xml
