+++ RPGCharacterSheet.MAUI/Platforms/Windows/App.xaml
