+++ RPGCharacterSheet.MAUI/Platforms/Windows/Package.appxmanifest
