+++ RPGCharacterSheet.MAUI/Properties/launchSettings.json
