+++ RPGCharacterSheet.MAUI/Resources/Raw/AboutAssets.txt
